# packages/contracts/openapi/backend-api.yaml
openapi: 3.1.0
info:
  title: NutriScan Backend API
  version: 0.1.0
  description: Contract for the Mobile App to Backend API boundary.
paths:
...
  /me/profile:
    get:
      summary: Get the current anonymous user's profile.
      responses:
        "501":
          description: Contract placeholder. Endpoint shape is not finalized yet.
    put:
      summary: Create or update the current anonymous user's profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileRequest"
      responses:
        "501":
          description: Contract placeholder. Endpoint shape is not finalized yet.
...
components:
  schemas:
...
    UpdateUserProfileRequest:
      type: object
      properties:
        heightCm:
          type: number
          minimum: 1
        weightKg:
          type: number
          minimum: 1
        ageRange:
          type: string
          enum:
            - under_18
            - 18_24
            - 25_34
            - 35_44
            - 45_54
            - 55_plus
      required:
        - heightCm
        - weightKg
...